pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bg-test-data"
version = "0.1.0"
description = "Comprehensive Bulgarian test data generator with valid checksums"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Nikolay Chillev" }]
keywords = ["bulgaria", "test-data", "egn", "bulstat", "iban", "faker", "qa", "testing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "allure-pytest>=2.13",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
bg-test-data = "bg_test_data.cli:main"

[project.urls]
Homepage = "https://github.com/Nikolay-Chillev/bg-test-data"
Repository = "https://github.com/Nikolay-Chillev/bg-test-data"
Issues = "https://github.com/Nikolay-Chillev/bg-test-data/issues"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "smoke: quick validation tests",
    "unit: unit tests for individual generators",
    "integration: tests verifying cross-module correlation",
]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.coverage.run]
source = ["bg_test_data"]
branch = true

[tool.coverage.report]
fail_under = 90
